refactor type structure
