+++ packages/las-sdk-core/src/types/secret.ts
@@ -1,0 +1,24 @@
+import { RequestConfig, PaginationOptions } from './common';
+
+export type Secret = {
+  secredId: string;
+  description: string | null;
+  name: string | null;
+};
+
+export type ListSecretsOptions = RequestConfig & PaginationOptions;
+
+export type SecretList = {
+  secrets: Array<Secret>;
+  nextToken: string | null;
+};
+
+export type CreateSecretOptions = RequestConfig & {
+  description?: string;
+};
+
+export type UpdateSecretOptions = RequestConfig & {
+  data?: Record<any, any>;
+  description?: string | null;
+  name?: string | null;
+};
